refactor: moved ReachSegment to ark-survey module/folder
