+++ src/app.module.ts
@@ -25,7 +25,6 @@ import { ObjectModule } from './schema/object/object.module';
 import { HealthController } from './HealthController';
 import { SurveyModule } from './schema/survey/survey.module';
 import { ArkSurveyModule } from './schema/ark-survey/ark-survey.module';
-import { ReachSegmentModule } from './schema/reach-segment/reach-segment.module';
 
 @Module({
 	imports: [
@@ -58,7 +57,6 @@ import { ReachSegmentModule } from './schema/reach-segment/reach-segment.module'
 		}),
 		TerminusModule,
 		ArkSurveyModule,
-		ReachSegmentModule,
 	],
 	controllers: [HealthController],
 	providers: [
