+++ src/schema/ark-survey/__stubs__/index.ts
@@ -1,1 +1,2 @@
 export * from './ark-survey-stub';
+export * from './reach-segment-stub';
