+++ src/schema/ark-survey/__stubs__/reach-segment-stub.ts
@@ -4,8 +4,8 @@ import { domainSurvey } from '../../survey/__stubs__';
 import { SurveyFactory } from '../../survey/survey.factory';
 import { ReachSegment as DomainReachSegment } from '../types/reach-segment.repository.interface';
 import { CreateReachSegmentInput } from '../dto/create-reach-segment.input';
-import { ReachSegmentFactory } from '../reach-segment.factory';
 import { UpdateReachSegmentInput } from '../dto/update-reach-segment.input';
+import { ReachSegmentFactory } from '../reach-segment.factory';
 
 domainSurvey.inspectionStandardType = 'quaywalls';
 export const arkSurveyStub = SurveyFactory.CreateSurvey(domainSurvey);
