+++ src/schema/ark-survey/commands/create-reach-segment.command.ts
