+++ src/schema/ark-survey/commands/find-ark-survey-reach-segments.handler.ts
@@ -1,10 +1,9 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
-import { ReachSegment } from '../../reach-segment/models/reach-segment.model';
-import { ReachSegmentService } from '../../reach-segment/reach-segment.service';
+import { ReachSegment } from '../models/reach-segment.model';
+import { ReachSegmentService } from '../reach-segment.service';
 
 import { FindArkSurveyReachSegmentsCommand } from './find-ark-survey-reach-segments.command';
-
 @CommandHandler(FindArkSurveyReachSegmentsCommand)
 export class FindArkSurveyReachSegmentsHandler implements ICommandHandler<FindArkSurveyReachSegmentsCommand> {
 	constructor(private service: ReachSegmentService) {}
