+++ src/schema/ark-survey/commands/find-ark-survey-reach-segments.spec.ts
@@ -1,11 +1,10 @@
 import { MockedObjectDeep } from 'ts-jest';
 
-import { ReachSegmentService } from '../../reach-segment/reach-segment.service';
-import { ReachSegment as reachSegmentStub } from '../../reach-segment/__stubs__';
+import { ReachSegmentService } from '../reach-segment.service';
+import { ReachSegment as reachSegmentStub } from '../__stubs__';
 
 import { FindArkSurveyReachSegmentsHandler } from './find-ark-survey-reach-segments.handler';
 import { FindArkSurveyReachSegmentsCommand } from './find-ark-survey-reach-segments.command';
-
 
 const reachSegmentsServiceMock: MockedObjectDeep<ReachSegmentService> = {
 	findReachSegments: jest.fn().mockResolvedValue([reachSegmentStub]),
