+++ src/schema/ark-survey/dto/base-reach-segment.input.ts
