+++ src/schema/ark-survey/models/ark-survey.model.ts
@@ -2,7 +2,7 @@ import { Field, Float, ObjectType } from '@nestjs/graphql';
 import { Point } from 'graphql-geojson-scalar-types';
 import { Point as PointType } from 'geojson';
 
-import { ReachSegment } from '../../reach-segment/models/reach-segment.model';
+import { ReachSegment } from './reach-segment.model';
 
 @ObjectType({ description: 'arkSurvey' })
 export class ArkSurvey {
