+++ src/schema/ark-survey/models/reach-segment.model.spec.ts
