+++ src/schema/ark-survey/models/reach-segment.model.ts
