+++ src/schema/ark-survey/reach-segment.service.ts
