+++ src/schema/ark-survey/types/reach-segment.repository.interface.ts
