--- src/schema/reach-segment/__stubs__/index.ts
@@ -1,1 +1,0 @@
-export * from './reach-segment-stub';
